Added files via upload
Just changed the var/www to the correct file path for use on Rpi 3b. It is now var/www/html. You will have to delete an index.html file from Apache in order for it to work. But once you do so it should work on the new pi without the need of an additional wifi dongle right out of the box. If I didn't do this correctly I'm sorry, first time editing here.